Start Radiant rebuild app shell
